research_dating/pyproject.toml
[project]
name = "research-dating"
version = "0.1.0"
description = "RAG-powered faculty-student research matching tool for BSPH"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.28.0",
    "voyageai>=0.2.0",
    "chromadb>=0.5.0",
    "pdfplumber>=0.11.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "tiktoken>=0.7.0",
    "nltk>=3.8.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "sentence-transformers>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "chunking", "vectordb", "pipeline", "baseline", "evaluation", "app"]
